[SPARK-14452][SQL] Provide ways to specify encoders in Scala without implicits
